walrus/requirements.txt
Django==2.2.3
djangorestframework==3.9.4
django[argon2]==2.0.1.10
flake8>=3.6.0, <=3.7.0
